% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{bench}
\alias{bench}
\title{Benchmark the overhead.}
\usage{
bench(n = 1e+06, k_numeric = 10, k_char = 10, k_date = 10, seed = 1234)
}
\arguments{
\item{n}{number of observation in the data}

\item{k_numeric}{number of numeric columns}

\item{k_char}{number of character columns}

\item{k_date}{number of date columns}

\item{seed}{seed for random number generation}
}
\value{
a list with benchmark results
}
\description{
Benchmark the overhead.
}
