% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{conda_create}
\alias{conda_create}
\title{Create conda environment.}
\usage{
conda_create(
  envname = NULL,
  packages = "python",
  conda = "auto",
  forge = TRUE,
  channel = character()
)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{packages}{A character vector, indicating package names which should be
installed.}

\item{conda}{The path to a `conda` executable. Use `"auto"` to allow
`reticulate` to automatically find an appropriate `conda` binary.}

\item{forge}{Boolean; include the [Conda Forge](https://conda-forge.org/)
repository?}

\item{channel}{An optional character vector of Conda channels to include.
When specified, the `forge` argument is ignored. If you need to
specify multiple channels, including the Conda Forge, you can use
`c("conda-forge", <other channels>)`.}
}
\value{
string with path to Python binary
}
\description{
This function is extending
https://github.com/rstudio/reticulate/blob/d30786a7ca1335003c52c458c1b4081d47ddc9bc/R/conda.R#L109
which is licensed under Apache License 2.0,
see https://github.com/rstudio/reticulate/blob/master/LICENSE
with copyright 2016-2017 RStudio, Inc.
}
