% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{create_cluster}
\alias{create_cluster}
\title{Create cluster that will be used to encapsulate reticulate sessions}
\usage{
create_cluster(force = FALSE)
}
\arguments{
\item{force}{logical indicating whether the cluster
shall be recreated if it exists}
}
\description{
Create cluster that will be used to encapsulate reticulate sessions
}
