% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{create_package_env}
\alias{create_package_env}
\title{Create conda environment that belongs to this package.}
\usage{
create_package_env()
}
\value{
Exit code from reticulate::conda_create() call.
}
\description{
Create conda environment that belongs to this package.
}
