% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{encapsulate}
\alias{encapsulate}
\title{Run function via makeCluster to allow several python bindings in reticulate}
\usage{
encapsulate(func, ...)
}
\arguments{
\item{func}{is the function that will be run}

\item{...}{are the function arguments}
}
\description{
Run function via makeCluster to allow several python bindings in reticulate
}
