% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{get_cluster_type}
\alias{get_cluster_type}
\title{Set cluster type for encapsulation.}
\usage{
get_cluster_type()
}
\value{
type cluster type used for makeCluster
}
\description{
Set cluster type for encapsulation.
}
