% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{get_conda_path}
\alias{get_conda_path}
\title{Get conda path that is used inside package}
\usage{
get_conda_path()
}
\value{
conda path (or "auto" if reticulate is supposed to
  automatically find a conda binary)
}
\description{
Get conda path that is used inside package
}
