% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{get_conda_requirements}
\alias{get_conda_requirements}
\title{Get conda requirements that are shipped with this package}
\usage{
get_conda_requirements()
}
\value{
vector with requirements
}
\description{
Get conda requirements that are shipped with this package
}
