% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{get_package_envname}
\alias{get_package_envname}
\title{Get name of conda environment that belongs to this package}
\usage{
get_package_envname()
}
\value{
vector with name of conda environment
}
\description{
Get name of conda environment that belongs to this package
}
