% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{package_env_exists}
\alias{package_env_exists}
\title{Check if the package environment exists}
\usage{
package_env_exists()
}
\value{
boolean indicating whether environment exists
}
\description{
Check if the package environment exists
}
