% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{python_model_predict}
\alias{python_model_predict}
\title{Predict method for Python Model}
\usage{
python_model_predict(df)
}
\arguments{
\item{df}{data frame to be passed on to Python predict function}
}
\value{
vector with predictions
}
\description{
This function passes an input data frame on to a Python function
and retrns a vector of predictions.
}
