% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{set_cluster_type}
\alias{set_cluster_type}
\title{Set cluster type for encapsulation.}
\usage{
set_cluster_type(type)
}
\arguments{
\item{type}{cluster type used for makeCluster}
}
\description{
Set cluster type for encapsulation.
}
