% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{set_conda_path}
\alias{set_conda_path}
\title{Set conda path to be used inside package}
\usage{
set_conda_path(conda)
}
\arguments{
\item{conda}{The path to a `conda` executable. Use `"auto"` to allow
`reticulate` to automatically find an appropriate `conda` binary.}
}
\description{
Set conda path to be used inside package
}
