% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{set_encapsulate}
\alias{set_encapsulate}
\title{Set encapsulate}
\usage{
set_encapsulate(encapsulate)
}
\arguments{
\item{encapsulate}{Logical indicating whether to encapsulate
reticulate calls}
}
\description{
Set encapsulate
}
