% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Close cluster that was be used to encapsulate reticulate sessions}
\usage{
stop_cluster()
}
\description{
Close cluster that was be used to encapsulate reticulate sessions
}
